test(onboard): split gateway runtime coverage
